recipe_scraper/requirements.txt
flask>=3.0
recipe-scrapers>=15.0
pywebview>=6.1
requests>=2.31
pillow>=10.0
